Lock new solvers behind regression coverage and lighter benchmark recording
This adds dedicated frontier regression coverage for the LTMOA-family
variants, extends the benchmark smoke matrix, and trims summary-only
benchmark overhead so canonical runs do not pay for trace details they
never export. Along the way it also fixes several noisy signedness and
unused-variable issues in existing code.

Constraint: Benchmark schema must stay stable while summary-only runs get cheaper
Rejected: Add new benchmark output schema for summary-only mode | would fragment existing analysis scripts
Rejected: Leave new solver behavior untested outside manual runs | too risky for repeated algorithm work
Confidence: high
Scope-risk: moderate
Reversibility: clean
Directive: Keep summary-only and trace-enabled benchmark modes both covered when changing recorder behavior
Tested: Built main, ltmoa_frontier_regression, benchmark_metrics_smoke; ran both tests and ctest in current workspace
Not-tested: Full benchmark sweep after this commit in isolation

+++ inc/algorithms/gcl/gcl_tree.h
@@ -13,7 +13,7 @@ public:
     using Snapshot = std::vector<CostVec<N>>;
 
     Gcl_tree(size_t num_node) {
-        for (int i = 0; i < num_node + 1; i++) {
+        for (size_t i = 0; i < num_node + 1; i++) {
             gcl.push_back(AVL_Tree<N>());
         }
     };
